% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{AnnotateClusters}
\alias{AnnotateClusters}
\title{Annotate Clusters with User-Defined Labels}
\usage{
AnnotateClusters(object, label_mapping)
}
\arguments{
\item{object}{A `SpatialExperiment` object.}

\item{label_mapping}{A named vector where names correspond to existing cluster
labels and values correspond to the new labels.}
}
\value{
The input object with a new `"LabelledCluster"` column in its `colData`.
}
\description{
This function assigns user-defined labels to clusters in a
`SpatialExperiment` object based on a mapping.
}
\examples{
# Annotate clusters
label_map <- c(
    "T cell_1" = "Treg",
    "T cell_2" = "CD4+ T cell",
    "T cell_3" = "CD8+ T cell"
)
spe <- AnnotateClusters(object = spe, label_mapping = label_map)
}
