% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{ClusterCells}
\alias{ClusterCells}
\title{Cluster Cells Based on Marker Expression or Dimensional Reduction}
\usage{
ClusterCells(
  object,
  k,
  prefix,
  markers = rownames(object),
  filter_condition = NULL,
  assay_name = "standardised",
  seed = 123,
  batch_corrected = FALSE
)
}
\arguments{
\item{object}{A `SpatialExperiment` object.}

\item{k}{An integer specifying the number of nearest neighbors to use for clustering.}

\item{prefix}{A character string prefix added to the front of cluster names
(e.g., `prefix = "T cell"` will produce clusters named `"T cell_1"`, `"T cell_2`, etc.)}

\item{markers}{A character vector of markers to use for clustering. Default
is all markers (`rownames(object)`).}

\item{filter_condition}{An optional condition for filtering cells prior to
clustering, provided as a tidy evaluation formula (e.g., `celltype == "T cell"`).}

\item{assay_name}{A character string specifying the assay to use for clustering if
`batch_corrected = FALSE`. Default is `"standardised"`.}

\item{seed}{An integer seed for reproducibility. Default is 123.}

\item{batch_corrected}{A logical value. If `TRUE`, the `"batch_corrected"`
reduced dimension is used instead of marker expression. Default is `FALSE`.}
}
\value{
The input object with an updated `"Cluster"` column in its `colData`.
  If cells are already assigned to a cluster, that cluster identity will be
  overwritten with the new clustering results.
}
\description{
This function performs clustering on cells in a `SpatialExperiment` object
using either marker expression or a batch-corrected reduced dimension.
}
\examples{
# Filter for T cells, perform clustering with 30 nearest neighbours, and name
# resulting clusters with the "T cell" prefix
spe <- ClusterCells(
    object = spe,
    k = 30,
    prefix = "T cell",
    filter_condition = celltype == "T cell"
)

# Perform clustering of B cells using 20 nearest neighbours and
# batch-corrected reduced dimension instead of assay data
spe <- ClusterCells(
    object = spe,
    k = 20,
    prefix = "B cell",
    filter_condition = celltype == "B cell",
    batch_corrected = TRUE
)
}
