% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CreateMantisFolder}
\alias{CreateMantisFolder}
\title{Create Mantis Project Folder with Data}
\usage{
CreateMantisFolder(
  object,
  img_suffix = "_full.tiff",
  mask_suffix = "_CpSeg_mask.tif",
  assay_name = "scale.data",
  analysis_path = "analysis",
  mantis_path = "analysis/5_R_analysis/MantisProject"
)
}
\arguments{
\item{object}{A `SpatialExperiment` object.}

\item{img_suffix}{A character string specifying the suffix added to the end of the full image names.
Default is `"_full.tiff"`.}

\item{mask_suffix}{A character string specifying the suffix added to the end of the mask image names.
Default is `"_CpSeg_mask.tif"`.}

\item{assay_name}{A character string specifying the assay to use for TIFF generation.
Default is `"scale.data"`.}

\item{analysis_path}{A character string specifying the path to analysis data.
Default is `"analysis"`.}

\item{mantis_path}{A character string specifying the path for the Mantis project.
Default is `"MantisProject"`.}
}
\value{
None. Files are written to the specified Mantis folder.
}
\description{
This function generates a folder structure compatible with Mantis, including
TIFF images, segmentation masks, and metadata.
}
\examples{
CreateMantisFolder(object = spe, raw_path = "raw", analysis_path = "analysis")
}
