% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{CreateSortedVector}
\alias{CreateSortedVector}
\title{Create Sorted Vector}
\usage{
CreateSortedVector(n)
}
\arguments{
\item{n}{An integer specifying the maximum value in the sequence to be sorted.}
}
\value{
A sorted integer vector from `1` to `n`.
}
\description{
This function generates a sorted vector of integers from `1` to `n`.
It sorts the integers as characters using the `radix` method for
consistent lexical ordering.
}
\examples{
CreateSortedVector(10)
}
