% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{CreateSpatialExperiment}
\alias{CreateSpatialExperiment}
\title{Create a SpatialExperiment Object}
\usage{
CreateSpatialExperiment(analysis_path = "analysis", raw_path = "raw")
}
\arguments{
\item{analysis_path}{A character string indicating the path to the analysis folder.
Default is "analysis".}

\item{raw_path}{A character string indicating the path to the raw data folder.
Default is "raw".}
}
\value{
A `SpatialExperiment` object with processed assays, metadata, and
  spatial coordinates.
}
\description{
This function creates a `SpatialExperiment` object by loading, filtering,
and processing image, cell, and panel data. It generates spatial coordinates,
and assigns colour palettes for visualizations.
}
\examples{
spe <- CreateSpatialExperiment(
  analysis_path = "path/to/analysis",
  raw_path = "path/to/raw"
)
}
