% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ExportToMantis}
\alias{ExportToMantis}
\title{Export Metadata to Mantis-Compatible CSV}
\usage{
ExportToMantis(object, column, file_path)
}
\arguments{
\item{object}{A `SpatialExperiment` object.}

\item{column}{A character string specifying the column in `colData` to export.}

\item{file_path}{A character string specifying the file path for the output CSV.}
}
\value{
None. A CSV file is written to the specified path.
}
\description{
This function extracts and exports metadata from a `SpatialExperiment` object
to a CSV file compatible with Mantis.
}
\examples{
ExportToMantis(object = spe, column = "Cluster", file_path = "mantis_export.csv")
}
