% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ImportFlowJoData}
\alias{ImportFlowJoData}
\title{Import Data from FlowJo CSV Files}
\usage{
ImportFlowJoData(object, FlowJo_path)
}
\arguments{
\item{object}{A `SpatialExperiment` object.}

\item{FlowJo_path}{A character string specifying the path to the FlowJo export directory.}
}
\value{
The input object with updated `FlowJo_celltype` annotations in its `colData`.
}
\description{
This function imports FlowJo annotations into a `SpatialExperiment` object
and updates the metadata.
}
\examples{
spe <- ImportFlowJoData(object = spe, FlowJo_path = "FlowJo")
}
