% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{NormaliseData}
\alias{NormaliseData}
\title{Normalise Data in a SpatialExperiment Object}
\usage{
NormaliseData(object)
}
\arguments{
\item{object}{A `SpatialExperiment` object.}
}
\value{
The input object with arcsinh-transformed data added as a new `"data"`
  assay, and z-standardised data added as a new `"standardised"` assay.
}
\description{
This function applies an arcsinh-transformation to the raw `"counts"` assay,
before standardising the data across channels within each image by
converting them to z-scores.
}
\examples{
spe <- NormaliseData(spe)
}
