% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReduction.R
\name{PerformPCA}
\alias{PerformPCA}
\title{Perform Principal Component Analysis (PCA)}
\usage{
PerformPCA(
  object,
  ncomponents,
  prefixes = NULL,
  markers = rownames(object),
  assay_name = "standardised",
  seed = 123,
  ...
)
}
\arguments{
\item{object}{A `SpatialExperiment` object.}

\item{ncomponents}{An integer specifying the number of principal components
to compute.}

\item{prefixes}{A character vector of prefixes used to subset the cells for
PCA. Default is `NULL` (all cells are used).}

\item{markers}{A character vector of marker names to include in the PCA.
Default is all markers (`rownames(object)`).}

\item{assay_name}{A character string specifying the assay to use for PCA.
Default is `"standardised"`.}

\item{seed}{An integer seed for reproducibility. Default is 123.}

\item{...}{Additional arguments passed to `scater::runPCA`.}
}
\value{
The input object with PCA results stored in reduced dimensions:
  - `"PCA_full"`: PCA on all cells.
  - `"PCA_subset"`: PCA on a subset of cells (if `prefixes` are provided).
}
\description{
This function performs PCA on a `SpatialExperiment` object.
PCA can be performed on all cells or a subset defined by prefixes.
}
\examples{
# Perform PCA on all cells
spe <- PerformPCA(object = spe, ncomponents = 10)

# Perform PCA on a subset of cells
spe <- PerformPCA(object = spe, ncomponents = 10, prefixes = c("B cell", "T cell"))
}
