% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotHeatmap}
\alias{PlotHeatmap}
\title{Plot Heatmap for Features}
\usage{
PlotHeatmap(
  object,
  prefix,
  annot_by,
  genes = rownames(object),
  assay_name = "scale.data",
  disp_min = NULL,
  disp_max = NULL,
  seed = 123,
  aggregate = FALSE,
  aggregate_by = NULL,
  ...
)
}
\arguments{
\item{object}{A `SpatialExperiment` object.}

\item{prefix}{A character string to filter clusters by prefix.}

\item{annot_by}{A column in `colData` used for annotation.}

\item{genes}{A character vector of gene names to include in the heatmap.
Default is all row names.}

\item{assay_name}{A character string specifying the assay to use for plotting.
Default is `"scale.data"`.}

\item{disp_min}{Minimum value for display. Values below will be clipped. Default is `NULL`.}

\item{disp_max}{Maximum value for display. Values above will be clipped. Default is `NULL`.}

\item{seed}{A numeric seed for reproducibility. Default is `123`.}

\item{aggregate}{Logical, whether to aggregate data. Default is `FALSE`.}

\item{aggregate_by}{A column in `colData` to use for aggregation. Required if `aggregate = TRUE`.}

\item{...}{Additional arguments passed to `dittoHeatmap`.}
}
\value{
A `pheatmap` object.
}
\description{
Creates a heatmap of feature expression from a `SpatialExperiment` object.
Allows aggregation and annotation.
}
\examples{
PlotHeatmap(object = spe, prefix = "Cluster", annot_by = "DonorID", genes = c("Gene1", "Gene2"))
}
