% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotRidge}
\alias{PlotRidge}
\title{Plot Ridge Plots for Features}
\usage{
PlotRidge(
  object,
  assay_name,
  features,
  split_by = NULL,
  ncol = 1,
  geom_density_ridges_args = list()
)
}
\arguments{
\item{object}{A `SpatialExperiment` object.}

\item{assay_name}{A character string specifying the assay to use for plotting.}

\item{features}{A character vector of feature names to plot.}

\item{split_by}{An optional column name in `colData` to split the ridges by.
Default is `NULL`.}

\item{ncol}{Number of columns for arranging plots in the grid. Default is `1`.}

\item{geom_density_ridges_args}{A list of additional arguments passed to `geom_density_ridges`.}
}
\value{
A `patchwork`ed `ggplot2` object.
}
\description{
Creates ridge plots for specified features from a `SpatialExperiment` object.
}
\examples{
PlotRidge(object = spe, assay_name = "scale.data", features = c("Gene1", "Gene2"))
}
