% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotScatter}
\alias{PlotScatter}
\title{Plot Scatter of Two Features}
\usage{
PlotScatter(
  object,
  feature1,
  feature2,
  assay_name1 = "scale.data",
  assay_name2 = "scale.data",
  colour_by = NULL,
  remove_na = TRUE
)
}
\arguments{
\item{object}{A `SpatialExperiment` object.}

\item{feature1}{A character string specifying the first feature (x-axis).}

\item{feature2}{A character string specifying the second feature (y-axis).}

\item{assay_name1}{A character string specifying the assay for `feature1`. Default is `"scale.data"`.}

\item{assay_name2}{A character string specifying the assay for `feature2`. Default is `"scale.data"`.}

\item{colour_by}{A column in `colData` to color points by. Default is `NULL`.}

\item{remove_na}{Logical, whether to remove cells with `NA` in `colour_by`. Default is `TRUE`.}
}
\value{
A `ggplot2` object.
}
\description{
Creates a scatter plot of two features from a `SpatialExperiment` object.
}
\examples{
PlotScatter(object = spe, feature1 = "Gene1", feature2 = "Gene2", colour_by = "Cluster")
}
