% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{UseMarkers}
\alias{UseMarkers}
\title{Retrieve a Set of Markers for Analysis}
\usage{
UseMarkers(colname = "Full")
}
\arguments{
\item{colname}{A character string specifying the column in panel.csv to filter by.
Default is "Full".}
}
\value{
A character vector containing the names of the selected markers.
}
\description{
This function extracts a set of markers from panel.csv based on a specified column.
The extracted markers can then be used as input for other functions in downstream analyses.
}
\examples{
# Extract markers from the "Full" column
markers <- UseMarkers(colname = "Full")

# Extract markers based on a different column
markers_subset <- UseMarkers(colname = "Subset")
}
