% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CreateMantisFolder}
\alias{CreateMantisFolder}
\title{Create Mantis Project Folder with Data}
\usage{
CreateMantisFolder(
  object,
  assay_name = "standardised",
  raw_path = "raw",
  analysis_path = "analysis",
  mantis_path = "MantisProject"
)
}
\arguments{
\item{object}{A `SpatialExperiment` object.}

\item{assay_name}{A character string specifying the assay to use for TIFF generation.
Default is `"standardised"`.}

\item{raw_path}{A character string specifying the path to raw data. Default is `"raw"`.}

\item{analysis_path}{A character string specifying the path to analysis data.
Default is `"analysis"`.}

\item{mantis_path}{A character string specifying the path for the Mantis project.
Default is `"MantisProject"`.}
}
\value{
None. Files are written to the specified Mantis folder.
}
\description{
This function generates a folder structure compatible with Mantis, including
TIFF images, segmentation masks, and metadata.
}
\examples{
CreateMantisFolder(object = spe, raw_path = "raw", analysis_path = "analysis")
}
